[OpenCL] Add HWLOC_OBJ_GROUP filter for Hwloc (#14585)
CMPLRLLVM-49622: On machines configured with SNC, a socket can be
devided into multiple NUMA nodes, but getProcessorIndexMap() can't get
the correct mapping of NUMA nodes to PUs. So HWLOC_OBJ_GROUP was added
to the hwloc, which enables hwloc insert objects, like PUs to NUMA nodes
according to their distances.