[ACD] Ensure, function definitions with forward declarations, carry auto-cpu-dispatch target-metadata (#14576)
This PR moves llvm.auto.cpu.dispatch and llvm.auto.arch target-metadata attachment from
SetFunctionAttributes() to SetLLVMFunctionAttributesForDefinition(). This is because the
metadata are attached only to function definitions in clang. To determine if a FunctionDecl
has a definition, existing implementation relied on the FunctionDecl::hasBody() member function.
However, it's been observed that FunctionDecl::hasBody() does not always return true for functions
that have bodies when called inside SetFunctionAttributes(), whereas, it does return true,
when called inside SetLLVMFunctionAttributesForDefinition(),.