[VPlan] Add support for reductions of fminimum/fmaximum intrinsics (#14621)
The community LoopVectorizer added a new recurrence kind for reductions
based on the llvm.fminimum and llvm.fmaximum intrinsics.  These are
IEEE-compliant versions of llvm.minnum and llvm.maxnum, which propagate
NaNs and treat -0.0 as less than +0.0.  This patch adds support in
VPlan for generating and handling the new recurrence kinds.

Reference: CMPLRLLVM-49037