[DTrans] Stop running the legacy DTrans transform passes that only work with typed pointers (#14671)
The compiler has switched to using opaque pointers, and the community
is in the process of removing support of typed pointers completely.

This is part 1 of the DTrans cleanup to stop putting the DTrans
transformation
passes that only worked with typed pointers into the pass pipeline
because the
code in the passes will stop compiling once the community removes some
of
the interfaces. When opaque pointers are in use, these passes all just
bail out
early after a check for the type of pointers in use. The LIT tests for
these passes
were removed long ago.

Also, updates one of the VPlan tests to use the replacement DTrans
transformation.

[CMPLRLLVM-49741]