Remove DTransAnalysis dependency from Intel_IndirectCallConv pass (#14768)
This is part 3 of the DTrans cleanup to remove the typed pointer DTrans
implementation in preparation of the community removing support for
typed pointers.

The DTransAnalysis pass only worked with typed pointers which are
going away, and no longer used by default. The Intel_IndirectCallConv
pass would uses the new DTransSafetyAnalyzer when opaque pointers
are in use. This removes the legacy path used for typed pointers.
LIT tests for this path were already removed during the change of
tests to opaque pointers.

[CMPLRLLVM-49741]