[VPlan] Avoid negative net costs due to unroll heuristic. (#14394)
Closes CMPLRLLVM-49253.

The unroll heuristic for partial sums had the potential to create
negative costs, as it's cost reduction was based on the total cost
associated with the reduction, as computed in analyze() based on the TTI
costs of instructions used only by the reduction.

This value is used to compute a ratio of (reduction_cost /
total_loop_cost), which can be > 1 - in the case from the Jira issue,
these costs are reduced due to OVLS loads. This ratio is now clamped to
1.

This value was also used to compute the cost adjustment due to partial
sums. This is too aggressive and subject to the same issue as the ratio
computation. This computation now uses the cost associated with the
reduction operation(s) alone (i.e. the binop(s) themselves).

Assertions are added:
* when the unroll heuristic computes a cost reduction exceeding the
total cost
* when the denominator of the estimated speedup in selectBestPlan() is 0
* when the net loop cost after heuristics is negative (for any reason)
in selectBestPlan()

The last set of assertions is based on @sguggill's comments on Jira
suggesting that costs should never be negative, which is the cause of
the potential division by zero.

A regression lit test has been added; for this I adapted
vplan_cost_model_ovls.ll to add a reduction, selecting VF/UF to
demonstrate the negative cost due to OVLS loads. Please let me know if I
should use any particular convention(s) to associate this to the Jira.