Remove DTransAnalysis dependency from DTransFieldModRef (#14773)
This is part 4 of the DTrans cleanup to remove the typed pointer DTrans
implementation in preparation of the community removing support for
typed pointers.

The DTransAnalysis pass only worked with typed pointers which are
going away, and no longer used by default. The pass would use the
DTransAnalysis pass when typed pointers were present, and the
DTransSafetyAnalyzer pass when opaque pointers are in use. This
change causes it to only use the DTransSafetyAnalyzer because
typed pointers are no longer being used.

This change resulted in the need to update one of the loop tests
that used the results of the DTransFieldModRef to add DTrans
type metadata to the test, which is required for the
DTransSafetyAnalyzer
pass to analyze the test.

[CMPLRLLVM-49741]