[VPlan] Add a VPlan-IR alignment propagation pass to VPlanDriver before CG. (#14604)
Add a method to VPlanAlignmentAnalysis which takes a vector VPlan and
guaranteed peel variant and propagates alignment to affected
load/stores.

This method is called on vector VPlans which are not peel loops, after
merger VPlans are created and before CG. and allows us to remove logic
in the LLVM path to check whether aligned load/stores need to be
emitted. This also fixes a bug where we would incorrectly use peeled
alignment for load/stores in vectorized peel loops.

JR: CMPLRLLVM-48910.