Resolve of merge (conflict) 3a45b843dec1 [lld] Preliminary fat-lto-object support 'main' -> 'xmain-web' (#14830)
CONFLICT (content): Merge conflict in lld/ELF/Driver.cpp

  commit 3a45b843dec1bca195884aa1c5bc56bd0e6755b4
  Author: Paul Kirth <paulkirth@google.com>
  Date:   Fri Mar 24 00:14:48 2023 +0000

      [lld] Preliminary fat-lto-object support

      This patch adds support to lld for --fat-lto-objects. We add a new
--fat-lto-objects flag to LLD, and slightly change how it chooses input
      files in the driver when the flag is set.

Fat LTO objects contain both LTO compatible IR, as well as generated
object
code. This allows users to defer the choice of whether to use LTO or not
to
link-time. This is a feature available in GCC for some time, and makes
the
existing -ffat-lto-objects flag functional in the same way as GCC's.

If the --fat-lto-objects option is passed to LLD and the input files are
fat
object files, then the linker will chose the LTO compatible bitcode
sections
embedded within the fat object and link them together using LTO.
Otherwise,
standard object file linking is done using the assembly section in the
object
      files.

Original RFC:
https://discourse.llvm.org/t/rfc-ffat-lto-objects-support/63977

      Depends on D146777

      Reviewed By: MaskRay

      Differential Revision: https://reviews.llvm.org/D146778

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Wenwan Xing <wenwan.xing@intel.com>