[VPlan] Remove nontyped SIMD clause tests (#14829)
The nontyped SIMD clause tests are no longer needed. Moved hir_simd_linear_3.ll test and 
changed it to use typed clauses as this test did not have counterpart. The test 
vplan_user_defined_reduction.ll was the same as vplan_user_defined_reduction_1.ll and for
other tests we had a test with the same name using typed clauses.