[llvm-profgen] Don't use duplication factors for IP-based profiles (#14833)
LBR-based execution profiles are aggregated by taking the maximum hit
count of multiple binary locations mapped to the same source location.
It doesn't make sense for instructions in a control flow path to have
different execution counts. To avoid under-counting duplicated source
code, e.g., unrolled loop bodies, a duplication factor is used to scale
up the aggregated maximum.

In the case of IP-only profiles, however, instructions in the same
control flow path may well have different sample counts. In order to
ensure that the profile reflects the total count for each debug location
aggregation is done by summing. In this case we don't want to apply the
duplication factor as it would result in over-counting.