Fix for CMPLRLLVM-46893 (#14779)
This patch moves the the SyncDependenceAnalysis files to our VPlan-
specific directory since the community has now deprecated it. The
algorithm still works as intended for vectorization and the new
UniformityAnalysis that the community now uses is not necessary for
us to adopt at this time. Note that we already had a set of
VPlanSyncDependenceAnalysis files, but this is set up this way to
enforce templates inside of the actual SyncDependenceAnalysis code,
so both sets of files are needed.