UnpredictableProfileLoader: annotate based on mispredict ratios (#14837)
Previously Intel_UnpredictableProfileLoader added !unpredictable
metadata based on the presence of some absolute number of branch
mispredict samples.

With this change it also loads the execution frequency profile
(specified by -fprofile-sample-use) and attempts to compute a mispredict
ratio. The !unpredictable metadata is only added for instructions with a
sufficiently high mispredict ratio.

This now requires some care that the two profiles have comparable
effective sampling rates, and this is complicated by the fact that
execution frequency profiles are created from LBR sampling, while
mispredict profiles are created by simpler IP sampling. The
`llvm-profgen` tool can be used to accomplish this:

    perf record -b -e br_inst_retired.near_taken:uppp,br_misp_retired.all_branches:upp -c 1000003 ...
    llvm-profgen --output freq.prof --sample-period=1000003 ...
    llvm-profgen --output misp.prof --sample-period=1000003 --leading-ip-only ...

Note: the --sample-period is implemented in #14835.