Bugfix to make AA treat dummy copy insts created by HIR as unknown ptrs (#14825)
HIR creates llvm.ssa.copy() LLVM intruction with undef operand to
represent
temps in HIR. If these temps are pointer types, they should be treated
as
unknown ptrs by AA so it doesn't incorrectly return noalias when dealing
with
them. This patch implements such behavior.

In addition, there is a small fix to treat llvm.ssa.copy() as always
preserving
nullness. This is needed to prevent BasicAA from treating it as an
'escape source'.

CMPLRLLVM-49211