Resolve of merge (conflict) 2dcf05125992 [CodeGen] Store call frame size in MachineBasicBlock 'main' -> 'xmain-web' (#15019)
CONFLICT (content): Merge conflict in
llvm/lib/CodeGen/MachineBasicBlock.cpp

  commit 2dcf0512599280ba68a5028f24ea96dfdb37b7b7
  Author: Jay Foad <jay.foad@amd.com>
  Date:   Fri Jun 30 08:14:40 2023 +0100

      [CodeGen] Store call frame size in MachineBasicBlock

Record the call frame size on entry to each basic block. This is usually
zero except when a basic block has been split in the middle of a call
      sequence.

This simplifies PEI::replaceFrameIndices which previously had to visit
      basic blocks in a specific order and had special handling for
unreachable blocks. More importantly it paves the way for an equally
simple implementation of a backwards version of replaceFrameIndices,
which is required to fully convert PrologEpilogInserter to backwards
register scavenging, which is preferred because it does not rely on
      accurate kill flags.

      Differential Revision: https://reviews.llvm.org/D156113

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: sys_ce_bb <sys_ce_bb@intel.com>