cherry pick vec abi unification (#14992)
Presently sycl::vec is not trivially or device copyable, and has a
different interface for device and host. This is not conforming with the
SYCL2020 spec. This was fixed recently on intel/llvm. Here we are cherry
picking this so as to get it into the next release. The
ABI-breaking-window is being extended for this patch.

With this change sycl::vec now uses std::array on host in all cases, but
device access is via ext_vector_type so as to not lose performance.
Alignment is set the same as the size to a maximum of 64 bytes, which
aligns with an upcoming spec change proposal. In the cases where the
size is greater than 64 we use std::array on the device. We no longer
need to emit a pragma on Windows and we can use alignas() on all
platforms.

https://jira.devtools.intel.com/browse/CMPLRLLVM-32626