Resolve of merge (conflict) 531aabfdee91 [SYCL] vec abi unification and trivially copyable (#9492) 'sycl' -> 'xmain-web' (3 commits) (#15102)
CONFLICT (content): Merge conflict in
sycl/source/detail/builtins_relational.cpp

  commit 531aabfdee91e0055e3d28459432dfa59b17850d
  Author: Chris Perkins <chris.perkins@intel.com>
  Date:   Tue Aug 1 17:07:32 2023 -0700

      [SYCL] vec abi unification and trivially copyable (#9492)

      adapted from https://github.com/intel/llvm/pull/6816

sycl::vec now uses std::array on host in all cases, but device access is
via `ext_vector_type` so as to not lose performance. Alignment is set
the same as the size to a maximum of 64 bytes, which aligns with an
upcoming spec change proposal. In the cases where the size is greater
than 64 we use std::array on the device. We no longer need to emit a
      pragma on Windows and we can use `alignas()` on all platforms.

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Jinsong Ji <jinsong.ji@intel.com>