Resolve of merge (conflict) cb3136b0d359 [clang] Pass --cuda-path to fix test/Driver/openmp-offload.. 'main' -> 'xmain-web' (#15174)
CONFLICT (content): Merge conflict in
clang/test/Driver/openmp-offload-jit.c

  commit cb3136b0d3596f5c3984e4fb49359e2a1a28a547
  Author: Jonas Hahnfeld <jonas.hahnfeld@cern.ch>
  Date:   Tue Aug 8 09:46:36 2023 +0200

      [clang] Pass --cuda-path to fix test/Driver/openmp-offload-jit.c

This test was trying to detect a system installation of CUDA and was
marked as returning exit code 1 as part of D156363. Pass an explicit
CUDA installation to make the test return exit code 0 regardless of
a CUDA being found on the system or not. Also add an explicit -march
      to get a stable test.

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Wenwan Xing <wenwan.xing@intel.com>