CMPLRLLVM-50298: Ensure that name is taken by new function when replaceFunctionWithFunction is called in inlining reports. (#15154)
The member function replaceFunctionWithFunction() was not placed
consistenly in the optimizations that needed it. It should be placed
after the old function has taken the name of the new function. This is
importnat in the metadata inlining report, for which temps are
controlled by a map using the name of the function aa a key.

I have added asserts and moved the code in ArgumentPromotion to be
consistent with the other cases, and added a new LIT test.