CMPLRLLVM-50359: Fix non-determinism in using FunctionNameIndexMap in metadata inlining report. (#15201)
FunctionNameIndexMap in the metadata inlining report uses a StringRef as
a
key to represent the function name. But this function name can be passed
from
one function to another using takeName, which is not accounted for my
this map.
Changing the key to a std::string and the map to a std::map fixes this.

I have also added an internal option to dump the FunctionNameIndexMap
before it is used compaction, to check for consistency.