[SYCL] Ignore vec_type_hint attribute in SYCL 2020 (#10619) (#15267)
According to the SYCL 2020 spec, [[sycl::vec_type_hint()]] attribute
should accept arguments of the type sycl::vec type. The attribute should
also be accepted with non conforming lambda syntax.

The current implementation in SYCL corresponds to the openCL version of
this argument (with an additional spelling for SYCL), i.e. the attribute
accepts extended vector type, floating point types and integral type. An
error diagnostic is thrown for sycl:vec type.

Since the attribute is deprecated and is not handled by any SYCL
backend, and will be removed in a future version of the spec, to be
minimally conformant with SYCL 2020 spec, this PR just ignores the
attribute instead of adding support for sycl::vec type. Support was also
added for non conforming lambda syntax