[HIRCreation] Link the latch exit of inner multi-exit loop at the correct loop level when it belongs to an ancestor loop (#15199) (#15229)
When the latch exit of inner multi-exit loop jumps to an ancestor loop,
it is
still linked in the parent loop. This was causing parsing to set
inconsistent
def@ levels for temps defined in this exit bblock. This PR fixes the
issue by
linking the bblock at the correct loop level.

CMPLRLLVM-50257

https://github.com/intel-restricted/applications.compilers.llvm-project/pull/15199