[SYCL] mark failing image tests as unsupported at O0.  (#10793) (#15367)
The spir-v generation is not working correctly for legacy images at O0.
There is no intention of fixing it at this time, given that the image
thing is going to be completely overhauled in the future. Marking these
tests as unsupported in O0.

cherry-pick PR [(#10793)](https://github.com/intel/llvm/pull/10793) into
xmain.