[SYCL][ESIMD] Add checks for zero-geps in simd object codegen test (#15420)
In xmain zero-geps are not being optimized away to aid TBAA info for
Dtrans, and for OpenMP pocessing. See details in CMPLRLLVM-36462.

It is not clear whether it is the best way to compile device code,
as these optimizations might not be relevant for device code, however
that needs to be checked. Perhaps, zero-gep optimization can be
turned on with -xmain-enable-gep0-removal flag for device code.