Resolve of merge (conflict) 14e3bec8fc3e Reland "[lld] Preliminary fat-lto-object support" 'main' -> 'xmain-web' (#15553)
CONFLICT (content): Merge conflict in lld/ELF/Driver.cpp

  commit 14e3bec8fc3e1f10c3dc57277ae3dbf9a4087b1c
  Author: Paul Kirth <paulkirth@google.com>
  Date:   Mon Jul 24 20:14:17 2023 +0000

      Reland "[lld] Preliminary fat-lto-object support"

      This patch adds support to lld for --fat-lto-objects. We add a new
--fat-lto-objects option to LLD, and slightly change how it chooses
input
      files in the driver when the option is set.

Fat LTO objects contain both LTO compatible IR, as well as generated
object
code. This allows users to defer the choice of whether to use LTO or not
to
link-time. This is a feature available in GCC for some time, and makes
the
existing -ffat-lto-objects option functional in the same way as GCC's.

If the --fat-lto-objects option is passed to LLD and the input files are
fat
object files, then the linker will chose the LTO compatible bitcode
sections
embedded within the fat object and link them together using LTO.
Otherwise,
standard object file linking is done using the assembly section in the
object
      files.

The previous version of this patch had a missing `REQUIRES: x86` line in
`fatlto.invalid.s`. Additionally, it was reported that this patch caused
      a test failure in `export-dynamic-symbols.s`, however,
      29112a994694baee070a2021e00f772f1913d214 disabled the
`export-dynamic-symbols.s` test on Windows due to a quotation difference
      between platforms, unrelated to this patch.

Original RFC:
https://discourse.llvm.org/t/rfc-ffat-lto-objects-support/63977

      Reviewed By: MaskRay

      Differential Revision: https://reviews.llvm.org/D146778

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Tyler Lanphear <tyler.lanphear@intel.com>