[BPI] Fix branch probability for weighted averaging pixels loop (#15427)
CMPLRLLVM-48247: The estimated trip count for hottest loop in this JIRA
is far less than real scenario. Thus, the loop block frequency is under
real value. This leads to RA split a register in this hot loop and
caused performance regression.
This patch scales up weight of this loop back edge by factor 3 to make
the block frequency closer to real value so that RA won’t split register
in this loop.
Also fixed an incorrect bracket when scaling likely blocks.