Update GitPython required version (#15494)
This change updates the required version of gitpython listed in the
llvm/utils/git/requirements.txt file.

We don't actually use this file in any of our development processes and
it isn't distributed with the product, but the version number here
triggers a CVE report from Dependabot. This file is used in the LLVM
open source release process, and it will be updated there. It is being
cherry-picked into xmain-rel to avoid problems with the security
process.

This addresses CMPLRLLVM-50583