[VPlan] Print location information for masked remainder "loops" (#15450)
When we produce a masked remainder loop during HIR code generation,
it generally has a known trip count of 1. In this case we call an HIR
utility to replace the loop with its first iteration.  This has a side effect of
preserving the optimization report for the no-longer-loop and attaching
it to the main loop as a sibling opt report. Sibling opt reports require
a debug location if we want to see it on the report, and VPlan is not
currently providing one.  For both code generators, make sure we
always provide a debug location on the opt report.

Reference:  CMPLRLLVM-49582