[SYCL][ESIMD] Add support for 512 GRF size specification to sycl_ext_intel_grf_size (#15533)
Falcon Shores supports 512 GRF size, and we want to allow user
specification of this GRF size through the existing `intel_grf_size`
experimental extension.

Extend the extension documentation to capture Falcon Shores. Note that
FS supports more than the three GRF sizes listed, but for now the
extension will only support these three.

Next, update the compile-time assert to allow 512, and update program
manager to pass the new flag to VC.

Note this new flag is not yet implemented, and it will be VC (so ESIMD)
only at first.

Once it is implemented we should be able to use this flag always instead
of `-doubleGRF`, at least for VC.

I added an E2E test that verifies the flag is passed to IGC, though the
executables currently error due to the flag not being implemented yet so
we expect an error. Once it is implemented, we can remove the `not` from
the E2E test RUN lines.

Finally, this is all under embargo.

Jira: CMPLRLLVM-50270

---------

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>