Resolve of merge (conflict) c5488c8dcc8c [OpenMP] Properly set static thread limit (w/o analysis) 'main' -> 'xmain-web' (#15691)
CONFLICT (content): Merge conflict in
clang/lib/CodeGen/CGOpenMPRuntime.cpp

  commit c5488c8dcc8cd005f4a04aec239b8d22afae51e3
  Author: Johannes Doerfert <johannes@jdoerfert.de>
  Date:   Sat Aug 19 12:51:20 2023 -0700

      [OpenMP] Properly set static thread limit (w/o analysis)

We used to have two separate implementations to derive the number of
threads used in a target region. This lead us to sometimes miss out on
      user provided thread bounds (num_threads, or thread_limit) when we
      looked for "constant default values". If we might miss out on the
presence of those bounds, we cannot set the thread_limit statically
since the runtime will try to honor user input rather than cap it at the
"preferred default". This patch replaces the secondary implementation
with the primary in a mode that will not emit code but just look for the
presence, and potentially upper bounds, of thread limiting clauses.

The runtime test would not pass without this rewrite as we missed some
clauses, set the static limit on the device to the preferred value, but
      then violated that value at runtime.

      Fixes: https://github.com/llvm/llvm-project/issues/64845

      Differential Revision: https://reviews.llvm.org/D158381

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: Tyler Lanphear <tyler.lanphear@intel.com>