[VPlan][SLP] Add commutativity basic support. (#15684)
The change implements support for 1 step ahead lookup when compose candidate
vectors out of operands of commutative operations. When the resulting vector
definitely can not be vectorized the operands are swapped in SLP worklist,
but not in VPlan.

Also the change relax requirements for splat vectors recognition
assuming that any VPValue can be broadcasted, as apposite VPInstruction requirement in
the code before the change.