Resolve of merge (conflict) aa78e94b0bc6 [Libomptarget] Support mapping indirect host calls to devi.. 'main' -> 'xmain-web' (#15808)
CONFLICT (content): Merge conflict in
openmp/libomptarget/include/omptarget.h
  CONFLICT (content): Merge conflict in openmp/libomptarget/src/rtl.cpp

  commit aa78e94b0bc66375de7f2383b4e39c07cd482104
  Author: Joseph Huber <jhuber6@vols.utk.edu>
  Date:   Mon Aug 14 16:19:29 2023 -0500

[Libomptarget] Support mapping indirect host calls to device functions

The changes in D157738 allowed for us to emit stub globals on the device
in the offloading entry section. These globals contain the addresses of
      device functions and allow us to map host functions to their
corresponding device equivalent. This patch provides the initial support
required to build a table on the device to lookup the associated value.
This is done by finding these entries and creating a global table on the
      device that can be searched with a simple binary search.

This requires an allocation, which supposedly should be automatically
freed at plugin shutdown. This includes a basic test which looks up
device
pointers via a host pointer using the added function. This will need to
be built
      upon to provide full support for these calls in the runtime.

To support reverse offloading it would also be useful to provide a
reverse table
      that allows us to get host functions from device stubs.

      Depends on D157738

      Reviewed By: jdoerfert

      Differential Revision: https://reviews.llvm.org/D157918

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: sys_ce_bb <sys_ce_bb@intel.com>
Co-authored-by: Wenwan Xing <wenwan.xing@intel.com>