CMPLRLLVM-49174: Fix for collecting the SIMD instructions when sibling loops are available (#15581)
This patch updates the process of collecting the HLInst that contains
DIR_OMP_SIMD (SIMD entry instruction) to consider if the SIMD end is
found early. This fixes the issue where a loop is SIMD, but its
sibling isn't and could be identified as SIMD.

Also, this patch updates the function getSIMDExitIntrinsic to use
the same process as getSIMDEntryIntrinsic but with the proper
directive.

This is the patch as PR #15497 and PR #15606, but for xmain-rel.