[VecClone] Fix for CMPLRLLVM-49470 (#15180) (#15829)
At -O0 we have difficulty processing linear uval args in VecClone
because we can't infer the pointer element type to generate the
stores to "fake" memory since the original stores are done indirectly.
The current fix bails out when we fail to detect the pointer element
type. See full problem description in Intel_VecClone.cpp. This fixes a
stability problem on the MFL, so we need to revisit and come up with
a more robust solution rather than bailing out.