[GlobalOpt] Optimize global variables that is only used by trivial load-store chains. (#13962)
For a global variable to be relevant in a program, it's value must be
used by
some instructions that affect the program's behavior. Some globals are
used by trivial load-store chains like:

  %1 = load i32, ptr @g
  %2 = add i32 %1, 1
  store i32 %2, ptr @g

The store is not constant, has side-effect but if the global is only
used in
such cases, it doesn't actually affect visible behavior. This patch
teaches
GlobalOpt to optimize this code.

JIRA: CMPLRLLVM-48720