[TLI][VPlan] Ignore FastMathFlags for vectorizing library calls when imf-use-svml option is used. (#15733)
We currently check if a FP type math library call has ApproxFunc fast
math flag to vectorize it with SVML version. However this check can be
ignored if user explicitly forced SVML usage in compiler via
"imf-use-svml" option. This PR updates the check in TLI which is used in
vectorizer.

JIRA: CMPLRTST-18906
Change-Id: I8460497a872933bff88fadecdf424ffd4503ee08