Resolve of merge (conflict) 5ae881ff0ab3 InstCombine: Fold out scale-if-denormal pattern 'main' -> 'xmain-web' (#16007)
CONFLICT (content): Merge conflict in
llvm/lib/Transforms/InstCombine/InstCombineSelect.cpp

  commit 5ae881ff0ab3a38121f240c9e73427367284e802
  Author: Matt Arsenault <Matthew.Arsenault@amd.com>
  Date:   Thu Aug 3 12:32:09 2023 -0400

      InstCombine: Fold out scale-if-denormal pattern

      Fold select (fcmp oeq x, 0), (fmul x, y), x => x

This cleans up a pattern left behind by denormal range checks under
      denormals are zero.

      The pattern starts out as something like:
        x = x < smallest_normal ? x * K : x;

The comparison folds to an == 0 when the denormal mode treats input
      denormals as zero. This makes library denormal checks free after
      linked into DAZ enabled code.

alive2 is mostly happy with this, but there are some issues. First,
there are many reported failures in some of the negative tests that
      happen to trigger some preexisting canonicalize introducing
combine. Second, alive2 is incorrectly asserting that denormals must
be flushed with the DAZ modes. It's allowed to drop a canonicalize.

      https://reviews.llvm.org/D157030

---------

Co-authored-by: iclsrc <ia.compiler.tools.git@intel.com>
Co-authored-by: sys_ce_bb <sys_ce_bb@intel.com>