[SYCL] … (#15984)
Update reduction_reducer_op_eq.cpp test to avoid dividing by z (#10771)

divisor may be 0 in `T Diff = (Expected / *Res) - T{1};`

----
This is a cherry-pick of Xing Wenwang's fix for a division-by-zero bug
in one of the reduction tests. This is needed for other incoming work to
xmain-rel. It is a test-only fix.

Co-authored-by: Wenwan Xing <wenwan.xing@intel.com>