[X86] Fix INTEL_SYCL_OPAQUEPOINTER_READY guard. (#16047)
After a47705f2d8659d38a1e494b51fe3fb361e891bda, the `xmain-web`
self-build started failing. This was due to a segfault in
help_txt_reader caused by a miscompilation: a load to an unaligned
address was being emitted as an aligned load, i.e:

    movdqa  _ZL8line_buf+6(%rip), %xmm0 ; <--- address is *not* aligned

This miscompilation was caused by the `expandmemcmp` pass, which expands
calls to `memcmp()` to a sequence of LLVM-IR instructions. During this
pass, an aligned load should only be emitted if the base ptr + offset is
itself aligned; however, an aligned load was *always* being emitted.
This was due to a bad `INTEL_SYCL_OPAQUEPOINTER_READY` guard added in
merge conflict #14205, which caused the alignment calculation to be
omitted when `-INTEL_DSYCL_POINTER_OPAQUEPOINTER=1` is passed.

This patch moves the guard so that the alignment calculation happens in
both paths. This causes the load to be correctly emitted as unaligned,
fixing the segfault in the self-build.

JR: CMPLRLLVM-51271